Add response to search queries
A new function was written to handle requests with get
parameters from searches on index or wiki pages. Now
similar results are shown for a given search i.e query
parameter.

This resolves "Search" requirements.

Official requirement:
Search: Allow the user to type a query into the search box in the
sidebar to search for an encyclopedia entry.
 - If the query matches the name of an encyclopedia entry, the user
   should be redirected to that entry’s page.
 - If the query does not match the name of an encyclopedia entry, the
   user should instead be taken to a search results page that displays a
   list of all encyclopedia entries that have the query as a substring.
   For example, if the search query were Py, then Python should appear
   in the search results.
 - Clicking on any of the entry names on the search results page should
   take the user to that entry’s page.