Add a page to create new entries
A page to enter new entries was created so there's
the possible to insert different wikis in the site.

This resolves "New Page" requirements

Official requirements:
Edit Page: On each entry page, the user should be able to click a link
to be taken to a page where the user can edit that entry’s Markdown
content in a textarea.
 - The textarea should be pre-populated with the existing Markdown
   content of the page. (i.e., the existing content should be the
   initial value of the textarea).
 - The user should be able to click a button to save the changes made to
   the entry.
 - Once the entry is saved, the user should be redirected back to that
   entry’s page.